Add new code parser and directory reader modules
