+++ src/backend/langflow/interface/custom/code_parser/__init__.py
@@ -1,0 +1,3 @@
+from .code_parser import CodeParser
+
+__all__ = ["CodeParser"]
