+++ src/backend/langflow/interface/custom/directory_reader/__init__.py
@@ -1,0 +1,3 @@
+from .directory_reader import DirectoryReader
+
+__all__ = ["DirectoryReader"]
