+++ src/backend/langflow/interface/custom/directory_reader/directory_reader.py
@@ -1,6 +1,7 @@
-import os
 import ast
+import os
 import zlib
+
 from loguru import logger
 
 
@@ -63,12 +64,13 @@ class DirectoryReader:
         return len(file_content.strip()) == 0
 
     def filter_loaded_components(self, data: dict, with_errors: bool) -> dict:
+        from langflow.interface.custom.utils import build_component
         items = [
             {
                 "name": menu["name"],
                 "path": menu["path"],
                 "components": [
-                    component
+                    (*build_component(component),component)
                     for component in menu["components"]
                     if (component["error"] if with_errors else not component["error"])
                 ],
